src/requirements.txt
apache-airflow-providers-snowflake==4.1.0
astronomer-cosmos==1.8.2